slug added to teacher create

+++ Tuition/Teacher/TeacherBAL.py
@@ -14,28 +14,33 @@ def is_teacher_exist(teacherid):
 
 def unlock_teacherBAL(user,teacherId):
     teacher = is_teacher_exist(teacherId)
-    if IsUserTeacherExist(user.id,teacher.id) == False and user.id != teacher.User_id.id :
+    if IsUserTeacherExist(user.id,teacher.id) == False and user.id != teacher.User_id.id and teacher is not None:
         if user.credit_points > 0 and  UnlockTeacher(user,teacher):
             cp = user.credit_points
             user.credit_points = cp -1
             user.save()
-            return True
+            teacher.phone_number
+            return teacher.phone_number
         else:
-            return False
+            return None
     return True
     
 
 
-def save_teacher(Name, Gender, Experience, 
-                 Location, Qualification, Subject, classes, 
-                 About, User_id, Teaching_mode, 
-                 Phone_number,Age,Fee,Pincode,photo=None):
-    return CreateTeacher(Name=Name, Gender=Gender,
-                        Experience=Experience, location=Location,
-                        Qualification=Qualification, Subject=Subject,
-                        classes=classes, About=About, User_id=User_id,
-                        Teaching_mode=Teaching_mode, Phone_number=Phone_number,
-                        Age=Age,Fee=Fee,Pincode=Pincode,photo=photo)
+def save_teacher(name, gender, experience, 
+                 location, qualification, subject, classes, 
+                 about, user_id, teaching_mode, 
+                 phone_number,age,fee,pincode,photo=None):
+    slug =pincode.Devision + pincode.District #getting distruc for slug from pincode object
+    slug = slug + '-' + classes.split()[0] + '-' + subject.split()[0] + '-' + location.split()[0]
+
+
+    return CreateTeacher(name=name, gender=gender,
+                        experience=experience, location=location,
+                        qualification=qualification, subject=subject,
+                        classes=classes, about=about, user_id=user_id,
+                        teaching_mode=teaching_mode, phone_number=phone_number,
+                        age=age,fee=fee,pincode=pincode,photo=photo,slug=slug)
 
 def get_latest_teacher(pageNumber):
     return getLatestTeacher(pageNumber)
